Add Food module, search, add and get endpoints

+++ src/main/java/com/dizio1/fittracker/food/dto/FoodNutrients.java
@@ -1,7 +1,7 @@
 package com.dizio1.fittracker.food.dto;
 
 public record FoodNutrients(
-        Integer nutrientId,
+        Long nutrientId,
         String nutrientName,
         String unitName,
         Double value
